Modifications
Modifications to the application code,

Fixed an issue that would cause the application to crash if there was no rain data present in the JSON.

Adding verification to the application when it is closed, this is to check with the user that they actually want to close down before actually closing.

Added the ability for the application to be run manually outside of the timer loop.

Added some verification to the export logs method to check that the settings had actually been set before allowing the export to go ahead.

Export logs no longer accepts a file name from the settings, instead this is hard coded into the application code and appended with the current date and time of the export.

Tidied up some of the labels that hadn't been named.